test common move
