+++ cpp/unittest/gis/cuda/CMakeLists.txt
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 set(gis_cuda_test_srcs
-        test_common/test_common.cpp
         transforms_tests.cpp
         functor_distance_tests.cpp
         functor_area_tests.cpp
