+++ cpp/unittest/gis/cuda/arrow_conversion_tests.cpp
@@ -26,7 +26,7 @@
 #include "gis/cuda/common/gis_definitions.h"
 #include "gis/cuda/functor/st_distance.h"
 #include "gis/cuda/functor/st_point.h"
-#include "gis/cuda/test_common/test_common.h"
+#include "gis/test_common/test_common.h"
 
 namespace arctern {
 namespace gis {
