+++ cpp/unittest/gis/cuda/transforms_tests.cpp
@@ -21,8 +21,8 @@
 
 #include "gis/cuda/common/gis_definitions.h"
 #include "gis/cuda/conversion/conversions.h"
-#include "gis/cuda/test_common/test_common.h"
 #include "gis/cuda/wkb/wkb_transforms.h"
+#include "gis/test_common/test_common.h"
 
 namespace arctern {
 namespace gis {
